fix: address CodeRabbit review comments
- $user.delete: use splice for in-place deletion (consistent with tasks)
- users/_index/+queries.server.ts: fix stale 'tasks' comment -> 'users'
- tasks/import.tsx: move {...form.props} before explicit Form attrs
- sign-in-2/index.tsx: throw await -> return redirectWithSuccess
- data-table-row-actions.tsx: text-red-500 -> text-destructive
- users/_index/index.tsx: simplify URLSearchParams, remove spread pattern,
  remove dead Tailwind classes (lg:flex-row, lg:space-y-0, lg:space-x-12)
- users-mutate-form.tsx: remove unnecessary optional chaining on fields.password
- Extract UserDeleteConfirmDialog to _index/+components/ to separate
  client component from server route module
